+++ es5/spell-config.js
@@ -6,26 +6,39 @@ var _fs = require('fs');
 
 var _fs2 = _interopRequireDefault(_fs);
 
+var _async = require('async');
+
+var _async2 = _interopRequireDefault(_async);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var fileLines = [];
 var globalDictionary = [];
 var fileDictionary = {};
-var isCrLf = false;
-var globalDictionaryIndex = -1;
+var sharedSpelling = {};
+var relativeSpelling = {};
 
-function parse() {
+function spellingFile(fileName) {
+  return {
+    fileName: fileName,
+    fileLines: [],
+    lastLineOfGlobalSpellings: -1,
+    isCrLf: false,
+    isDirty: false
+  };
+}
+
+function parse(spelling) {
   var lastNonCommentIndex = -1;
   var inGlobal = true;
   var currentFile = void 0;
-  fileLines.forEach(function (line, index) {
+  spelling.fileLines.forEach(function (line, index) {
     if (!line || line.indexOf('#') === 0) {
       return;
     }
     var fileMatch = line.match(/^\s*-\s+(.*)/);
     if (fileMatch) {
       if (inGlobal) {
-        globalDictionaryIndex = lastNonCommentIndex === -1 ? index : lastNonCommentIndex + 1;
+        spelling.lastLineOfGlobalSpellings = lastNonCommentIndex === -1 ? index : lastNonCommentIndex + 1;
         inGlobal = false;
       } else {
         fileDictionary[currentFile].index = lastNonCommentIndex + 1;
@@ -43,53 +56,83 @@ function parse() {
     lastNonCommentIndex = index;
   });
   // make sure we end on a new-line
-  if (fileLines[fileLines.length - 1]) {
-    fileLines[fileLines.length] = "";
+  if (spelling.fileLines[spelling.fileLines.length - 1]) {
+    spelling.fileLines[spelling.fileLines.length] = "";
   }
   if (inGlobal) {
-    globalDictionaryIndex = lastNonCommentIndex === -1 ? fileLines.length - 1 : lastNonCommentIndex + 1;
+    spelling.lastLineOfGlobalSpellings = lastNonCommentIndex === -1 ? spelling.fileLines.length - 1 : lastNonCommentIndex + 1;
   } else {
     fileDictionary[currentFile].index = lastNonCommentIndex;
   }
 }
 
-function emptyFile() {
-  fileLines = ["# markdown-spellcheck spelling configuration file", "# Format - lines beginning # are comments", "# global dictionary is at the start, file overrides afterwards", "# one word per line, to define a file override use ' - filename'", "# where filename is relative to this configuration file", ""];
-  globalDictionaryIndex = fileLines.length - 1;
+function emptyFile(spelling) {
+  spelling.fileLines = ["# markdown-spellcheck spelling configuration file", "# Format - lines beginning # are comments", "# global dictionary is at the start, file overrides afterwards", "# one word per line, to define a file override use ' - filename'", "# where filename is relative to this configuration file", ""];
+  spelling.lastLineOfGlobalSpellings = spelling.fileLines.length - 1;
+}
+
+function initConfig() {
+  globalDictionary = [];
+  fileDictionary = {};
+  sharedSpelling = spellingFile("./.spelling");
+  relativeSpelling = spellingFile("");
+}
+
+function loadAndParseSpelling(spelling, next) {
+  _fs2.default.readFile(spelling.fileName, { encoding: 'utf-8' }, function (err, data) {
+    if (err) {
+      emptyFile(spelling);
+      return next();
+    }
+    if (data.indexOf('\r') >= 0) {
+      spelling.isCrLf = true;
+      data = data.replace(/\r/g, "");
+    }
+
+    spelling.fileLines = data.split('\n');
+    parse(spelling);
+    return next();
+  });
 }
 
 function initialise(filename, done) {
-  _fs2.default.readFile(filename, { encoding: 'utf-8' }, function (err, data) {
-    if (err) {
-      emptyFile();
-      return done();
-    }
-    if (data.indexOf('\r') >= 0) {
-      isCrLf = true;
-      data = data.replace(/\r/g, "");
-    }
-    fileLines = data.split('\n');
-    parse();
+  initConfig();
+  relativeSpelling.fileName = filename;
+  var sharedSpellingOnly = filename === "./.spelling";
+  _async2.default.parallel([function (next) {
+    loadAndParseSpelling(sharedSpelling, next);
+  }, function (next) {
+    sharedSpellingOnly && next() || !sharedSpellingOnly && loadAndParseSpelling(relativeSpelling, next);
+  }], function () {
     return done();
   });
 }
 
-function writeFile(done) {
-  var data = fileLines.join(isCrLf ? "\r\n" : "\n");
-  _fs2.default.writeFile('./.spelling', data, function (err) {
-    if (err) {
-      console.error("Failed to save spelling file");
-      console.error(err);
-      process.exitCode = 1;
-    }
+function writeFile(done, relative) {
+  var spelling = relative ? relativeSpelling : sharedSpelling;
+  if (spelling.isDirty) {
+    var data = spelling.fileLines.join(spelling.isCrLf ? "\r\n" : "\n");
+    _fs2.default.writeFile(spelling.fileName, data, function (err) {
+      if (err) {
+        console.error("Failed to save spelling file");
+        console.error(err);
+        process.exitCode = 1;
+      } else {
+        spelling.isDirty = false;
+      }
+      done();
+    });
+  } else {
     done();
-  });
+  }
 }
 
-function addToGlobalDictionary(word) {
+function addToGlobalDictionary(word, relative) {
+  var spelling = relative ? relativeSpelling : sharedSpelling;
   globalDictionary.push(word);
-  fileLines.splice(globalDictionaryIndex, 0, word);
-  globalDictionaryIndex++;
+  spelling.fileLines.splice(spelling.lastLineOfGlobalSpellings, 0, word);
+  spelling.isDirty = true;
+  spelling.lastLineOfGlobalSpellings++;
   for (var filename in fileDictionary) {
     if (fileDictionary.hasOwnProperty(filename)) {
       fileDictionary[filename].index++;
@@ -97,10 +140,12 @@ function addToGlobalDictionary(word) {
   }
 }
 
-function addToFileDictionary(filename, word) {
+function addToFileDictionary(filename, word, relative) {
+  var spelling = relative ? relativeSpelling : sharedSpelling;
   if (fileDictionary.hasOwnProperty(filename)) {
     var fileDict = fileDictionary[filename];
-    fileLines.splice(fileDict.index, 0, word);
+    spelling.fileLines.splice(fileDict.index, 0, word);
+    spelling.isDirty = true;
     for (var dictionaryFilename in fileDictionary) {
       if (fileDictionary.hasOwnProperty(dictionaryFilename) && fileDictionary[dictionaryFilename].index >= fileDict.index) {
         fileDictionary[dictionaryFilename].index++;
@@ -108,9 +153,10 @@ function addToFileDictionary(filename, word) {
     }
     fileDict.words.push(word);
   } else {
-    fileLines.splice(fileLines.length - 1, 0, " - " + filename, word);
+    spelling.fileLines.splice(spelling.fileLines.length - 1, 0, " - " + filename, word);
+    spelling.isDirty = true;
     fileDictionary[filename] = {
-      index: fileLines.length - 1,
+      index: spelling.fileLines.length - 1,
       words: [word]
     };
   }